+++ extensions/sanity/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Sanity Changelog
 
+## [Major Update] - 2023-08-17
+
+- Replace manual token with CLI authentication
+- New features:
+  - Copy project ID
+  - Viewing members and tokens
+  - List datasets
+  - List and add CORS origins
+- Add cache for faster navigation
+
 ## [Initial Version] - 2023-08-10
 
 - Support searching for Sanity projects.